requirements.txt
allure-behave==2.12.0
allure-combine==1.0.8
allure-python-commons==2.12.0
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
beautifulsoup4==4.11.2
behave==1.2.6
behave-html-formatter==0.9.6
behave2cucumber==1.0.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.15.1
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
delayed-assert==0.3.2
EasyProcess==0.3
envyaml==1.10.211231
exitstatus==2.0.1
Flask==1.1.2
Flask-JWT-Extended==3.25.0
future==0.18.2
globals==0.1.0.post23
greenlet==2.0.2
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kazoo==2.9.0
MarkupSafe==1.1.1
more-itertools==8.6.0
mss==5.1.0
ncs==0.1.0
netmiko==3.3.3
ntc-templates==1.6.0
packaging==20.7
paramiko==2.7.2
parse==1.18.0
parse-type==0.5.2
pluggy==0.13.1
psycopg2==2.9.6
py==1.9.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyscopg2==66.0.2
pyserial==3.5
pysftp==0.2.9
pytest==6.1.2
python-helper==0.1.0.post2
PyYAML==5.3.1
reportportal-behave-client==1.1.1
reportportal-client==5.0.3
requests==2.25.1
scp==0.13.3
selenium==3.141.0
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.9
stcrestclient==1.9.0
tenacity==6.2.0
textfsm==1.1.0
toml==0.10.2
typing_extensions==4.5.0
urllib3==1.16.0
webdriver-manager==4.0.1
Werkzeug==1.0.1
xvfbwrapper==0.2.9
